- Move test helpers into dev/ directory - Delete test output directory after every test - Add test output directories to gitignore - Add npm-debug.log to gitignore template
