+++ .gitignore
@@ -8,6 +8,8 @@ Cargo.lock
 **/artifacts.json
 cli/lib
 create-neon/dist
+create-neon/create-neon-test-project
+create-neon/create-neon-manual-test-project
 test/cli/lib
 npm-debug.log
 rls*.log
