+++ create-neon/data/templates/.gitignore.hbs
@@ -2,3 +2,4 @@ target
 index.node
 **/node_modules
 **/.DS_Store
+npm-debug.log*
