+++ create-neon/tsconfig.json
@@ -21,6 +21,7 @@
     },
     "include": [
         "src/**/*",
+        "dev/**/*",
         "test/**/*"
     ],
     "exclude": [
